refactor: move sentry functionality to core package. Integrate sentry error with logger
