+++ packages/core/package.json
@@ -60,6 +60,7 @@
   "dependencies": {
     "@internxt/sdk": "^1.15.10",
     "@phosphor-icons/react": "2.0.9",
+    "@sentry/electron": "^7.11.0",
     "@tanstack/react-virtual": "^3.13.12",
     "check-disk-space": "^3.4.0",
     "electron-log": "^5.4.1",
