+++ src/apps/backups/process-files/replace-files.ts
@@ -1,7 +1,6 @@
 import { stat } from 'node:fs/promises';
 import { BackupsContext } from '@/apps/backups/BackupInfo';
 import { FilesDiff } from '@/apps/backups/diff/calculate-files-diff';
-import { captureSentryUploadError } from '@/apps/shared/sentry/sentry';
 import { Sync } from '@/backend/features/sync';
 import { scheduleRequest } from '../schedule-request';
 
@@ -18,17 +17,12 @@ export async function replaceFiles({ ctx, modified }: Props) {
       try {
         await scheduleRequest({ ctx, path, fn: () => Sync.Actions.replaceFile({ ctx, path, uuid: remote.uuid }) });
       } catch (error) {
-        ctx.logger.error({ msg: 'Error replacing file', path, error });
-
         const fileStats = await stat(path).catch(() => null);
 
-        await captureSentryUploadError({
-          error,
-          fileUuid: remote.uuid,
-          fileSize: fileStats?.size ?? 0,
-          sourcePath: path,
-          uploadSource: 'backup-upload',
-        });
+        ctx.logger.sentryError(
+          { msg: 'Error replacing file', path, error },
+          { fileUuid: remote.uuid, fileSize: fileStats?.size ?? 0, sourcePath: path, uploadSource: 'backup-upload' },
+        );
       }
     }),
   );
