+++ src/backend/features/backups/download/download-file.ts
@@ -1,7 +1,6 @@
 import { logger } from '@internxt/drive-desktop-core/build/backend';
 import { createWriteStream } from 'node:fs';
 import { ExtendedDriveFile } from '@/apps/main/database/entities/DriveFile';
-import { captureSentryDownloadError } from '@/apps/shared/sentry/sentry';
 import { pipeline } from '@/core/utils/pipeline';
 import { InxtJs } from '@/infra';
 import { LocalSync } from '../..';
@@ -39,15 +38,15 @@ export async function downloadFile({ file, contentsDownloader }: Props) {
     throw error;
   } catch (error) {
     LocalSync.SyncState.addItem({ action: 'DOWNLOAD_ERROR', path });
-    logger.error({ tag: 'BACKUPS', msg: 'Error downloading file', path, error });
-
-    await captureSentryDownloadError({
-      error,
-      fileUuid: file.uuid,
-      contentsId: file.contentsId,
-      fileSize: file.size,
-      destinationPath: file.absolutePath,
-      downloadFailureSource: 'backup-download',
-    });
+    logger.sentryError(
+      { tag: 'BACKUPS', msg: 'Error downloading file', path, error },
+      {
+        fileUuid: file.uuid,
+        contentsId: file.contentsId,
+        fileSize: file.size,
+        destinationPath: file.absolutePath,
+        downloadFailureSource: 'backup-download',
+      },
+    );
   }
 }
