+++ src/backend/features/sync/actions/services/upload-file.ts
@@ -1,7 +1,6 @@
 import { AbsolutePath, SyncModule } from '@internxt/drive-desktop-core/build/backend';
 import { stat } from 'node:fs/promises';
 import { addSyncIssue } from '@/apps/main/background-processes/issues';
-import { captureSentryUploadError } from '@/apps/shared/sentry/sentry';
 import { CommonContext } from '@/apps/sync-engine/config';
 import { isBottleneckStop } from '@/infra/drive-server-wip/in/helpers/error-helpers';
 import { environmentFileUpload } from '@/infra/inxt-js/file-uploader/environment-file-uploader';
@@ -40,13 +39,10 @@ export async function uploadFile({ ctx, path }: Props) {
   } catch (error) {
     if (isBottleneckStop({ error })) return;
 
-    await captureSentryUploadError({
-      error,
-      fileUuid: '',
-      fileSize: size,
-      sourcePath: path,
-      uploadSource: 'sync-upload',
-    });
+    ctx.logger.sentryError(
+      { msg: 'Error uploading file', path },
+      { error, fileUuid: '', fileSize: size, sourcePath: path, uploadSource: 'sync-upload' },
+    );
 
     throw error;
   }
